refactor: page파일에서 generateMetadata함수를 사용하기 위해 클라이언트 코드는 PageClient파일로 분리
